Fix to DF clustering engine
The DF clustering engine was 'switching' which IDs were considered cluster IDs and which weren't
Provided you were clustering a document set to itself this was fine.
If you had different 'documentid' spaces the some of the axis matches were being
lost for all models that use the SJoin processing model.
The 'visible' symtom is that Euclidean could give the wrong result